Working in the browser!
Refactored a bunch into the assets folder; mostly for syntax highlighting on html and js.
Crowbook needed a temporary directory. I set it to / and the zip implementation (or crowbook) seems to clean up after itself.
So far, it seems to work!
Still need to test it on my ereader, and test the share functionality on the phone.